+++ changelog.md
@@ -72,6 +72,7 @@
 - Web API:
   - #142 Take back in stock notifications into account.
   - #805 Add endpoints to assign discounts to entities.
+  - #821 Add endpoints for RecurringPayment and RecurringPaymentHistory.
 - Import:
   - #390 Add a setting for whether to send the completion email.
   - #377 Import cross- and checkout-selling products.